requirements.txt
SpeechRecognition==3.10.1
pyttsx3==2.90
pyaudio==0.2.14
openai==0.28.1
google-genai>=1.0.0
python-dotenv==1.0.0
requests==2.31.0
vosk>=0.3.45
python-telegram-bot>=21.0
# Neural voice (human-like girl voice)
edge-tts>=6.1.0
pygame>=2.5.0
# NLP (natural language processing)
spacy>=3.7.0
# Avatar (anime character floating on screen)
Pillow>=10.0.0
websockets>=12.0
beautifulsoup4==4.12.3
